functions/requirements.txt
# Phoenix Functions - Python Dependencies

# Azure Functions
azure-functions==1.18.0
azure-functions-worker==1.0.0

# Azure SDK
azure-cosmos==4.5.1
azure-eventhub==5.11.4
azure-identity==1.15.0
azure-monitor-query==1.2.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-web==7.2.0
azure-mgmt-monitor==6.0.2
azure-communication-email==1.0.0

# OpenAI
openai==1.6.1

# Kubernetes
kubernetes==28.1.0

# Data Science & ML
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.0.3

# HTTP & Async
aiohttp==3.9.1
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pydantic==2.5.2
dataclasses-json==0.6.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

